Enrichment modularisation (https://github.com/sequana/sequana/issues/652)
