+++ .github/workflows/main.yml
@@ -29,7 +29,7 @@ jobs:
     - name: Install dependencies conda
       run: |
         conda install -c conda-forge mamba --quiet
-        mamba install -c conda-forge -c bioconda --quiet -y kraken2 dsrc shustring samtools snpeff
+        mamba install -c conda-forge -c bioconda --quiet -y kraken2 dsrc shustring samtools snpeff bwa 
         conda list | grep ruamel
         conda update ruamel_yaml
     - name: Install sequana with pip
@@ -42,4 +42,7 @@ jobs:
         pytest --cov-report term --cov=sequana
     - name: coveralls
       run: |
+        pip install coveralls
         coveralls --service=github
+      env:
+          GITHUB_TOKEN: ${{ github.token }}
