+++ sequana/enrichment/__init__.py
@@ -1,0 +1,4 @@
+from .kegg import KEGGPathwayEnrichment
+from .panther import PantherEnrichment
+from .gsea import GSEA
+from .mart import Mart
