+++ sequana/enrichment/gsea.py
@@ -1,0 +1,51 @@
+# -*- coding: utf-8 -*-
+#
+#  This file is part of Sequana software
+#
+#  Copyright (c) 2016-2020 - Sequana Development Team
+#
+#  File author(s):
+#      Thomas Cokelaer <thomas.cokelaer@pasteur.fr>
+#      Etienne Kornobis <etienne.kornobis@pasteur.fr>
+#
+#  Distributed under the terms of the 3-clause BSD license.
+#  The full license is in the LICENSE file, distributed with this software.
+#
+#  website: https://github.com/sequana/sequana
+#  documentation: http://sequana.readthedocs.io
+#
+##############################################################################
+
+from pathlib import Path
+import os
+import json
+
+import gseapy
+
+
+import colorlog
+
+logger = colorlog.getLogger(__name__)
+
+
+__all__ = ["GSEA"]
+
+
+
+class GSEA:
+    def __init__(self, species):
+        pass
+
+    def enrichment(self, gene_list, verbose=False, background=None):
+        enr = gseapy.enrichr(
+            gene_list=gene_list,
+            gene_sets=self.gene_sets,
+            verbose=verbose,
+            background=background,
+            outdir="test",
+            no_plot=True,
+        )
+
+        return enr
+
+
