+++ sequana/enrichment/mart.py
@@ -1,8 +1,7 @@
-# -*- coding: utf-8 -*-
 #
 #  This file is part of Sequana software
 #
-#  Copyright (c) 2016-2020 - Sequana Development Team
+#  Copyright (c) 2016-2021 - Sequana Development Team
 #
 #  Distributed under the terms of the 3-clause BSD license.
 #  The full license is in the LICENSE file, distributed with this software.
@@ -68,9 +67,10 @@ class Mart:  # pragma: no cover
 
     dataset = property(_get_dataset, _set_dataset)
 
-    def query(self,
-        attributes=["ensembl_gene_id", "go_id", "entrezgene_id", "external_gene_name"]):
-
+    def query(
+        self,
+        attributes=["ensembl_gene_id", "go_id", "entrezgene_id", "external_gene_name"],
+    ):
         logger.info("Please wait. This may take a while depending on your connection")
         self.biomart.new_query()
         self.biomart.add_dataset_to_xml(self.dataset)
