+++ setup.cfg
@@ -7,7 +7,7 @@ all_files  = 1
 
 [tool:pytest]
 # do not use --cov because it interfers with travis command
-addopts= --durations=10 --verbose --cov sequana --cov-report term-missing --maxfail=4 --durations 10
+addopts= --durations=10 --verbose --cov sequana --cov-report term-missing --maxfail=10 --timeout 20
 filterwarnings= 
     ignore:tostring.*is deprecated
     ignore:zmq.* is deprecated
