+++ test/enrichment/__init__.py
@@ -1,0 +1,3 @@
+import os
+test_dir = os.path.dirname(os.path.realpath(__file__))
+
