+++ test/test_bamtools.py
@@ -6,8 +6,6 @@ from easydev import TempFile
 import pytest
 import os
 
-skiptravis = pytest.mark.skipif("TRAVIS_PYTHON_VERSION" in os.environ, reason="On travis")
-
 
 def test_is_sam_bam():
 
@@ -17,7 +15,6 @@ def test_is_sam_bam():
     datatest = sequana_data("test_measles.bam", "testing")
     assert is_bam(datatest) is True
 
-@skiptravis
 def test_is_cram():
     datatest = sequana_data("test_measles.cram", "testing")
     assert is_cram(datatest) is True
@@ -81,7 +78,7 @@ def test_bam(tmpdir):
     except:
         assert True
 
-@skiptravis
+@pytest.mark.xfail(reason="cram flaky")
 def test_cram():
     datatest = sequana_data("test_measles.cram", "testing")
     s = CRAM(datatest)
@@ -150,7 +147,6 @@ def test_cs_in_bam():
 def test_insert_size():
 
     d1 = sequana_data("test_measles.sam", "testing")
-    #d2 = sequana_data("test_measles.cram", "testing")
     d3 = sequana_data("test.bam", "testing")
     d4 = sequana_data("test_CS_tiny.bam")
